backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.10.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
pdfplumber==0.10.3
tabula-py==2.8.2
boto3==1.34.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
python-dotenv==1.0.0
sentry-sdk[fastapi]==1.38.0
fuzzywuzzy==0.18.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
tenacity==8.2.3
slowapi==0.1.9

# Note: testing dependencies moved to requirements-dev.txt

# Optional: Secret management backends (install as needed)
# boto3>=1.34.0  # For AWS Secrets Manager (already included above)
# hvac==2.3.0    # For HashiCorp Vault
